pyproject.toml
[project]
name = "dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "discord>=2.3.2",
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]
